restructure | move controllers & utils under new ui package
