--- src/main/java/david/buzas/whattoeat/controllers/MainController.java
@@ -1,6 +1,0 @@
-package david.buzas.whattoeat.controllers;
-
-
-public class MainController {
-
-}
