+++ src/main/java/david/buzas/whattoeat/ui/controllers/ConsumptionHistoryPageController.java
@@ -1,12 +1,12 @@
-package david.buzas.whattoeat.controllers;
+package david.buzas.whattoeat.ui.controllers;
 
 import david.buzas.whattoeat.WhatToEatModel;
 import david.buzas.whattoeat.WhatToEatApplication;
 import david.buzas.whattoeat.entities.Meal;
 import david.buzas.whattoeat.entities.MealConsumption;
 import david.buzas.whattoeat.repositories.Repository;
-import david.buzas.whattoeat.utils.DeleteConfirmationDialog;
-import david.buzas.whattoeat.utils.ErrorAlert;
+import david.buzas.whattoeat.ui.components.DeleteConfirmationDialog;
+import david.buzas.whattoeat.ui.components.ErrorAlert;
 import javafx.fxml.FXML;
 import javafx.scene.control.*;
 import javafx.util.StringConverter;
