+++ src/main/java/david/buzas/whattoeat/ui/controllers/MainController.java
@@ -1,0 +1,6 @@
+package david.buzas.whattoeat.ui.controllers;
+
+
+public class MainController {
+
+}
